+++ packages/ui/src/components/backgrounds/index.ts
@@ -1,3 +1,4 @@
 export * from './WarpBackground';
 export * from './FaultyTerminal';
 export * from './OrbBackground';
+export * from './InfinityAnim';